compiletest

Module common

Structs

Config

Configuration for compiletest per invocation.

TargetCfg

TargetCfgs

TestPaths

Enums

CodegenBackend

CompareMode

Debugger

Endian

FailMode

PanicStrategy

PassMode

RunFailMode

RunResult

Sanitizer

TestMode

TestSuite

Constants

KNOWN_CRATE_TYPES

KNOWN_TARGET_HAS_ATOMIC_WIDTHS

Known widths of target_has_atomic.

UI_COVERAGE

UI_COVERAGE_MAP

UI_EXTENSIONS

UI_FIXED

UI_RUN_STDERR

UI_RUN_STDOUT

UI_STDERR

UI_STDERR_16

UI_STDERR_32

UI_STDERR_64

UI_STDOUT

UI_SVG

UI_WINDOWS_SVG

Functions

builtin_cfg_names 🔒

default_os 🔒

default_reloc_model 🔒

expected_output_path

Used by ui tests to generate things like foo.stderr from foo.rs.

incremental_dir

Absolute path to the directory to use for incremental compilation. Example: /path/to/build/host-tuple/test/ui/relative/testname.mode/testname.inc

output_base_dir

Absolute path to the directory where all output for the given test/revision should reside. Example: /path/to/build/host-tuple/test/ui/relative/testname.revision.mode/

output_base_name

Absolute path to the base filename used as output for the given test/revision. Example: /path/to/build/host-tuple/test/ui/relative/testname.revision.mode/testname

output_relative_path

Absolute path to the directory where all output for all tests in the given relative_dir group should reside. Example:

output_testname_unique

Generates a unique name for the test, such as testname.revision.mode.

query_rustc_output 🔒

serde_parse_u32 🔒

supported_crate_types 🔒