compiletest
Module debuggers

Functions

analyze_cdb ğŸ”’

Returns Path to CDB

analyze_gdb ğŸ”’

Returns (Path to GDB, GDB Version)

configure_cdb ğŸ”’

configure_gdb ğŸ”’

configure_lldb ğŸ”’

extract_cdb_version ğŸ”’

extract_gdb_version ğŸ”’

extract_lldb_version ğŸ”’

Returns LLDB version

find_cdb ğŸ”’

is_android_gdb_target ğŸ”’

Returns true if the given target is an Android target for the purposes of GDB testing.

is_pc_windows_msvc_target ğŸ”’

Returns true if the given target is a MSVC target for the purposes of CDB testing.

not_a_digit ğŸ”’