compiletest::directives

Module directive_names

Constants

KNOWN_DIRECTIVE_NAMES 🔒

This was originally generated by collecting directives from ui tests and then extracting their directive names. This is not an exhaustive list of all possible directives. Instead, this is a best-effort approximation for diagnostics. Add new directives to this list when needed.

KNOWN_HTMLDOCCK_DIRECTIVE_NAMES 🔒

KNOWN_JSONDOCCK_DIRECTIVE_NAMES 🔒