compiletest::executor

Function filter_tests

fn filter_tests(opts: &Config, tests: Vec<CollectedTest>) -> Vec<CollectedTest>

Applies command-line arguments for filtering/skipping tests by name.

Adapted from filter_tests in libtest.

FIXME(#139660): After the libtest dependency is removed, redesign the whole filtering system to do a better job of understanding and filtering paths, instead of being tied to libtest’s substring/exact matching behaviour.