compiletest
Function make_test

pub(crate) fn make_test(
    cx: &TestCollectorCx,
    collector: &mut TestCollector,
    testpaths: &TestPaths,
)

For a single test file, creates one or more test structures (one per revision) that can be handed over to libtest to run, possibly in parallel.