compiletest
Module header

Modules

cfg 🔒

directives 🔒

needs 🔒

Structs

CheckDirectiveResult 🔒

EarlyProps

Properties which must be known very early, before actually running the test.

HeaderLine 🔒

The broken-down contents of a line containing a test header directive, which iter_header passes to its callback function.

HeadersCache

TestProps

Enums

IgnoreDecision 🔒

Constants

KNOWN_DIRECTIVE_NAMES 🔒

This was originally generated by collecting directives from ui tests and then extracting their directive names. This is not an exhaustive list of all possible directives. Instead, this is a best-effort approximation for diagnostics. Add new headers to this list when needed.

KNOWN_HTMLDOCCK_DIRECTIVE_NAMES 🔒

KNOWN_JSONDOCCK_DIRECTIVE_NAMES 🔒

Functions

check_directive 🔒

expand_variables 🔒

extract_llvm_version

extract_llvm_version_from_binary

extract_version_range 🔒

Takes a directive of the form "<version1> [- <version2>]", returns the numeric representation of <version1> and <version2> as tuple: (<version1> as u32, <version2> as u32).

ignore_cdb 🔒

ignore_gdb 🔒

ignore_lldb 🔒

ignore_llvm 🔒

iter_header 🔒

line_directive

Extract an (Option<line_revision>, directive) directive from a line if comment is present.

llvm_has_libzstd

For tests using the needs-llvm-zstd directive:

make_test_description

parse_normalize_rule 🔒

Parses the regex and replacement values of a //@ normalize-* header, in the format: