compiletest::header
Struct TestProps

pub struct TestProps {

    pub error_patterns: Vec<String>,
    pub regex_error_patterns: Vec<String>,
    pub compile_flags: Vec<String>,
    pub run_flags: Vec<String>,
    pub doc_flags: Vec<String>,
    pub pp_exact: Option<PathBuf>,
    pub(crate) aux: AuxProps,
    pub rustc_env: Vec<(String, String)>,
    pub unset_rustc_env: Vec<String>,
    pub exec_env: Vec<(String, String)>,
    pub unset_exec_env: Vec<String>,
    pub build_aux_docs: bool,
    pub unique_doc_out_dir: bool,
    pub force_host: bool,
    pub check_stdout: bool,
    pub check_run_results: bool,
    pub dont_check_compiler_stdout: bool,
    pub dont_check_compiler_stderr: bool,
    pub no_prefer_dynamic: bool,
    pub pretty_mode: String,
    pub pretty_compare_only: bool,
    pub forbid_output: Vec<String>,
    pub revisions: Vec<String>,
    pub incremental_dir: Option<PathBuf>,
    pub incremental: bool,
    pub known_bug: bool,
    pass_mode: Option<PassMode>,
    ignore_pass: bool,
    pub fail_mode: Option<FailMode>,
    pub check_test_line_numbers_match: bool,
    pub normalize_stdout: Vec<(String, String)>,
    pub normalize_stderr: Vec<(String, String)>,
    pub failure_status: Option<i32>,
    pub dont_check_failure_status: bool,
    pub run_rustfix: bool,
    pub rustfix_only_machine_applicable: bool,
    pub assembly_output: Option<String>,
    pub should_ice: bool,
    pub stderr_per_bitwidth: bool,
    pub mir_unit_test: Option<String>,
    pub remap_src_base: bool,
    pub llvm_cov_flags: Vec<String>,
    pub filecheck_flags: Vec<String>,
    pub no_auto_check_cfg: bool,
    pub has_enzyme: bool,
    pub add_core_stubs: bool,

}

Fields
error_patterns: Vec<String>regex_error_patterns: Vec<String>compile_flags: Vec<String>run_flags: Vec<String>doc_flags: Vec<String>

Extra flags to pass to rustdoc but not the compiler.

pp_exact: Option<PathBuf>aux: AuxProps

Auxiliary crates that should be built and made available to this test.

rustc_env: Vec<(String, String)>unset_rustc_env: Vec<String>exec_env: Vec<(String, String)>unset_exec_env: Vec<String>build_aux_docs: boolunique_doc_out_dir: bool

Build the documentation for each crate in a unique output directory. Uses <root output directory>/docs/<test name>/doc.

force_host: boolcheck_stdout: boolcheck_run_results: booldont_check_compiler_stdout: booldont_check_compiler_stderr: boolno_prefer_dynamic: boolpretty_mode: Stringpretty_compare_only: boolforbid_output: Vec<String>revisions: Vec<String>incremental_dir: Option<PathBuf>incremental: boolknown_bug: boolpass_mode: Option<PassMode>ignore_pass: boolfail_mode: Option<FailMode>check_test_line_numbers_match: boolnormalize_stdout: Vec<(String, String)>normalize_stderr: Vec<(String, String)>failure_status: Option<i32>dont_check_failure_status: boolrun_rustfix: boolrustfix_only_machine_applicable: boolassembly_output: Option<String>should_ice: boolstderr_per_bitwidth: boolmir_unit_test: Option<String>remap_src_base: boolllvm_cov_flags: Vec<String>

Extra flags to pass to llvm-cov when producing coverage reports. Only used by the “coverage-run” test mode.

filecheck_flags: Vec<String>

Extra flags to pass to LLVM’s filecheck tool, in tests that use it.

no_auto_check_cfg: bool

Don’t automatically insert any --check-cfg args

has_enzyme: bool

Run tests which require enzyme being build

add_core_stubs: bool

Build and use minicore as core stub for no_core tests in cross-compilation scenarios that don’t otherwise want/need -Z build-std.

Implementations

impl TestProps

pub fn new() -> Self

pub fn from_aux_file( &self, testfile: &Path, revision: Option<&str>, config: &Config, ) -> Self

pub fn from_file( testfile: &Path, revision: Option<&str>, config: &Config, ) -> Self

fn load_from( &mut self, testfile: &Path, test_revision: Option<&str>, config: &Config, )

Loads properties from testfile into props. If a property is tied to a particular revision foo (indicated by writing //@[foo]), then the property is ignored unless test_revision is Some("foo").

fn update_fail_mode(&mut self, ln: &str, config: &Config)

fn update_pass_mode( &mut self, ln: &str, revision: Option<&str>, config: &Config, )

pub fn pass_mode(&self, config: &Config) -> Option<PassMode>

pub fn local_pass_mode(&self) -> Option<PassMode>

pub fn update_add_core_stubs(&mut self, ln: &str, config: &Config)