compiletest::runtest::debugger
Struct DebuggerCommands

pub(super) struct DebuggerCommands {
    pub commands: Vec<String>,
    pub breakpoint_lines: Vec<usize>,
    check_lines: Vec<(usize, String)>,
    file: PathBuf,
}

Representation of information to invoke a debugger and check its output

Fields
commands: Vec<String>

Commands for the debuuger

breakpoint_lines: Vec<usize>

Lines to insert breakpoints at

check_lines: Vec<(usize, String)>

Contains the source line number to check and the line itself

file: PathBuf

Source file name

Implementations

impl DebuggerCommands

pub fn parse_from( file: &Path, config: &Config, debugger_prefixes: &[&str], ) -> Result<Self, String>

pub fn check_output(&self, debugger_run_result: &ProcRes) -> Result<(), String>

Given debugger output and lines to check, ensure that every line is contained in the debugger output. The check lines need to be found in order, but there can be extra lines between.