Crate compiletest

Modules

common

compute_diff

errors

header

json ğŸ”’

These structs are a subset of the ones found in rustc_errors::json.

raise_fd_limit ğŸ”’

read2 ğŸ”’

runtest

util

Structs

Stamp ğŸ”’

Functions

analyze_cdb ğŸ”’

Returns Path to CDB

analyze_gdb ğŸ”’

Returns (Path to GDB, GDB Version)

check_overlapping_tests ğŸ”’

collect_tests_from_dir ğŸ”’

common_inputs_stamp ğŸ”’

Returns a stamp constructed from input files common to all test cases.

configure_cdb ğŸ”’

configure_gdb ğŸ”’

configure_lldb ğŸ”’

extract_cdb_version ğŸ”’

extract_gdb_version ğŸ”’

extract_lldb_version ğŸ”’

Returns LLDB version

files_related_to_test ğŸ”’

find_cdb ğŸ”’

is_android_gdb_target ğŸ”’

Returns true if the given target is an Android target for the purposes of GDB testing.

is_pc_windows_msvc_target ğŸ”’

Returns true if the given target is a MSVC target for the purposes of CDB testing.

is_test

Returns true if file_name looks like a proper test file name.

is_up_to_date ğŸ”’

log_config

make_test ğŸ”’

make_test_closure ğŸ”’

make_test_name ğŸ”’

make_tests

modified_tests ğŸ”’

not_a_digit ğŸ”’

opt_str

opt_str2

parse_config

run_tests

stamp ğŸ”’

test_opts