compiletest
Module util

Macros

static_regex 🔒

Traits

PathBufExt

Functions

add_dylib_path

Adds a list of lookup paths to cmd’s dynamic library lookup path. If the dylib_path_var is already set for this cmd, the old value will be overwritten!

copy_dir_all

dylib_env_var

The name of the environment variable that holds dynamic library locations.

lib_path_env_var

logv

make_new_path

path_div 🔒